refactor(assigment): move logic to hook
